Always set solid redirect_url to /
Because we use a client id document, we have a fixed list of allowed
redirect urls.
ldo/solid-react sets the redirect url to the "current page" which means
if we log in from a page that isn't /, the provider will raise an error
because the redirect urls don't match what's in the client id document.
I guess this isn't an issue in the case of dynamic registration.
inrupt docs suggest that the redirect url should be stable anyway:
https://docs.inrupt.com/developer-tools/javascript/client-libraries/tutorial/authenticate-browser/#authenticate-browser-login-process
> Specify a static redirectUrl value that does not change with application routes or hash or query parameters.