Auth flow: Show loader when checking auth state, redirect correctly
During auth we know when the app is in the process of redirecting to the
login provider and back to get an auth key. In this case show a loading
spinner instead of an incorrect "you must be logged in" message.

Due to the default implementation of ldo/solid-react, we can't do the
same type of "redirect to where you were" logic that we used when just
using the inrupt solid-authn library. Instead track the changing url and
manually call navigate in an effect