use mo:derived_from for audio triple
We decided to use this predicate to say that the mp3 file was derived
from the midi performance.

We temporarily rolled back the change when restoring the demo videos, as
these previously use available_as and we wanted to get the demo workin.
While it's not strictly correct we decided to modify the demo data files
to also use derived_from when pointing to video recordings of the
performance in order to be consistent with midi performances