Remove inpainting elements
This is an svg element which floated outside of the bounds of the error
ribbon and was supposed to highlight on top of notes on the score.
It was never finished, and the svg was mis-aligned, causing it to
prevent mouse clicks on other parts of the score. Remove until we
implement it properly